security: Pass xattrs allocated by LSMs to the inode_init_security hook
In preparation for moving EVM to the LSM infrastructure, this patch adds
the full array of xattrs allocated by LSMs as a new parameter of the
inode_init_security hook. It will be used by EVM to calculate the HMAC on
all xattrs.

This solution has been preferred to directly replacing *name, *value and
*len with the full array, as LSMs would have had to scan it to find an
empty slot.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>