ima: Align ima_post_path_mknod() definition with LSM infrastructure
Change ima_post_path_mknod() definition, so that it can be registered as
implementation of the path_post_mknod hook.

Also, make sure that ima_post_path_mknod() is executed only if
(mode & S_IFMT) is equal to zero or S_IFREG.

While the check is not necessary now, it will be when ima_post_path_mknod()
is registered as implementation for the path_post_mknod hook. That hook
(to be introduced) will be executed for all cases after the switch() in
do_mknodat(), if there is no previous error.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>