security: Pass xattrs allocated by LSMs to the inode_init_security hook
In preparation for moving EVM to the LSM infrastructure, this patch
replaces the name, value, len triple with the xattr array pointer provided
by security_inode_init_security(). LSMs are expected to call the new
function lsm_find_xattr_slot() to find the first unused slot of the array
where the xattr should be written.

Boundary checking is being performed in the new function to ensure that
callers don't obtain a slot outside the allocated xattr array. It is
assumed that each LSM requests at most one xattr (this limitation can be
removed in the future). The new variable inode_init_security_hooks has been
introduced to determine at LSM registration time how many
inode_init_security hook implementations will be invoked.

This patch modifies also SELinux and Smack to search for an unused slot, to
have a consistent behavior across LSMs (the unmodified version would
overwrite the xattr set by the first LSM in the chain). It is also
desirable to have the modification in those LSMs, as they are likely used
as a reference for the development of new LSMs.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>