security: Pass xattrs allocated by LSMs to the inode_init_security hook
In preparation for moving EVM to the LSM infrastructure, this patch
replaces the name, value, len triple with the xattr array pointer provided
by security_inode_init_security().

LSMs are expected to call the new function lsm_find_xattr_slot() to find
the first unused slot of the array where the xattr should be written and
to update the xattr array argument to the pointer of the slot they found,
so that security_inode_init_security() can perform validation.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>