ima: Add template fields evmxattrnames, evmxattrlengths and evmxattrvalues
This patch introduces the new template fields evmxattrnames,
evmxattrlengths and evmxattrvalues, which contain respectively the
concatenated names (strings, separated by |), lengths (u32, hex) and
values (hex) of EVM protected xattrs. If an xattr is not present, the name
and length are not displayed in the measurement list.

Protected xattr values, together with the inode UID, GID and mode would
allow to verify the EVM portable signature, if it was included with the
template field sig or evmsig.

Protected xattr names and lengths, although useful to infer more
information about a file, are not required for verifying the EVM portable
signature.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>