bpf-preload: Export and call bpf_obj_do_pin_kernel()
Export bpf_obj_do_pin_kernel(), so that the preload method of an eBPF
program (built-in or in a kernel module) can call this function to pin
links, maps and progs.

This removes the need of defining the bpf_preload_info structure, to hold
the links to be pinned by populate_bpffs().

Since after this patch the preload method has everything needed to pin
objects, that makes it possible to pin arbitrary objects of eBPF programs,
by generating the corresponding preload method in the light skeleton.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>