bpftool: Generate code of kernel module - static vars
As the preloading code of a generic eBPF program is deterministic,
generate it from the eBPF object, similarly to the light skeleton. The
syntax of the new subcommand is:

bpftool gen module <eBPF object> <light skeleton> [ name <object name> ]

Since the generation code is long, split in multiple parts and include
here the main code of the kernel module and the static variables of the
objects to be pinned.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>