bpf: Introduce bpf_mod_verify_sig() helper
Introduce the bpf_mod_verify_sig() helper, which similarly to
mod_verify_sig() verifies a module-style appended signature.

Its main benefit is to let eBPF programs rely on the kernel trust anchor,
the primary and secondary keyrings, for the verification of the
authenticity of data, not only kernel modules.

The helper returns the size of data without the appended signature, so that
eBPF program can eventually parse the right amount of data.

In addition to mod_verify_sig(), bpf_mod_verify_sig() is also able to
verify PGP signatures.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>