bpf: Add bpf_map_same() helper
The security hook bpf_map could be used to control access to a specific
map. The mode parameter can be used to distinguish map reads from map
writes.

However, even 'bpftool map list' results in requesting write permission. If
that permission is not granted, bpftool stops iterating over the defined
maps.

The bpf security hook is instead much more accurate, it lets implementers
of the hook to make a decision only when the map is updated (with the
command BPF_MAP_UPDATE_ELEM or BPF_MAP_DELETE_ELEM).

However, unlike bpf_map, from an eBPF program it is not possible to compare
the address of a map with one the eBPF program cares about, as the only
other available information from the bpf() system call is the file
descriptor of the map.

Introduce the bpf_map_same() helper, which internally retrieves the map
associated to the file descriptor, and compares it with the address of a
map passed as argument.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>